refactor: relocated files
relocated files to the todo
changed by adding features
