+++ todo/src/main/java/com/ssafy/sandbox/todo/controller/TodoController.java
@@ -1,8 +1,8 @@
-package com.ssafy.sandbox.controller;
+package com.ssafy.sandbox.todo.controller;
 
-import com.ssafy.sandbox.model.domain.TodoDTO;
-import com.ssafy.sandbox.model.domain.TodoVO;
-import com.ssafy.sandbox.service.TodoService;
+import com.ssafy.sandbox.todo.model.domain.TodoDTO;
+import com.ssafy.sandbox.todo.model.domain.TodoVO;
+import com.ssafy.sandbox.todo.service.TodoService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
