+++ todo/src/main/java/com/ssafy/sandbox/todo/model/domain/TodoVO.java
@@ -1,4 +1,4 @@
-package com.ssafy.sandbox.model.domain;
+package com.ssafy.sandbox.todo.model.domain;
 
 public interface TodoVO {
     Long getId();
