+++ todo/src/main/resources/application-local.yml
@@ -21,4 +21,4 @@ spring:
         format_sql: true
 logging:
   level:
-    com.ssafy.todo: ${LOGGING_LEVEL:DEBUG}
+    com.ssafy.sandbox: ${LOGGING_LEVEL:DEBUG}
